Move components to folders
