+++ planner/src/components/plan/bottom_segments/BSNewModal.vue
