+++ planner/src/components/plan/bottom_segments/BottomSegments.vue
@@ -122,11 +122,11 @@
 <script lang="ts">
     import {Component, Vue} from "vue-property-decorator";
     import {namespace} from 'vuex-class';
-    import BSEditModal from "@/components/plan/BSEditModal.vue";
+    import BSEditModal from "@/components/plan/bottom_segments/BSEditModal.vue";
     import {segmentType} from "@/common/serde/segment_type";
-    import BSReorderModal from "@/components/plan/BSReorderModal.vue";
+    import BSReorderModal from "@/components/plan/bottom_segments/BSReorderModal.vue";
     import {displayBottomSegmentElement} from "@/common/display";
-    import BSNewModal from "@/components/plan/BSNewModal.vue";
+    import BSNewModal from "@/components/plan/bottom_segments/BSNewModal.vue";
     import {BottomSegmentElement} from "@/store/plan";
 
     const plan = namespace('Plan');
