+++ planner/src/components/plan/deco_gases/DecoGases.vue
